+++ vector.go
@@ -19,7 +19,7 @@ func Vec[T Number](x, y T) Vector[T] {
 // Transform creates a new Vector by applying the given matrix to the current vector.
 // For integer T, the float64 result of each component is rounded; rotations and
 // non-integer scales lose precision.
-func (v Vector[T]) Transform(matrix Matrix) Vector[T] {
+func (v Vector[T]) Transform(matrix Matrix[float64]) Vector[T] {
 	return Vector[T]{Cast[T](matrix.A*float64(v.X) + matrix.B*float64(v.Y)), Cast[T](matrix.D*float64(v.X) + matrix.E*float64(v.Y))}
 }